refactor: update screen file name
